+++ torchsynth/torchmodule.py
@@ -9,7 +9,7 @@ import torch
 import torch.nn as nn
 import torch.tensor as T
 
-import torchsynth.torchutil as util
+import torchsynth.util as util
 from torchsynth.defaults import BUFFER_SIZE, SAMPLE_RATE
 from torchsynth.parameter import ParameterRange, TorchParameter
